src/test/java/org/openshift/examples/eap/services/HelloWorldServiceTest.java
package org.openshift.examples.eap.services;

import org.jboss.arquillian.container.test.api.Deployment;
import org.jboss.arquillian.junit.Arquillian;
import org.jboss.arquillian.test.api.ArquillianResource;
import org.jboss.shrinkwrap.api.ShrinkWrap;
import org.jboss.shrinkwrap.api.asset.EmptyAsset;
import org.jboss.shrinkwrap.api.spec.JavaArchive;
import org.junit.Test;
import org.junit.runner.RunWith;

import javax.inject.Inject;
import java.io.*;
import java.net.HttpURLConnection;
import java.net.URL;
import java.nio.charset.Charset;

import static org.junit.Assert.assertEquals;

/**
 * Created by faisalmasood on 4/03/2016.
 */
@RunWith(Arquillian.class)
public class HelloWorldServiceTest {

    @Deployment
    public static JavaArchive createDeployment() {
        return ShrinkWrap.create(JavaArchive.class)
                .addClass(HelloWorldService.class)
                .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
    }

    @Inject
    HelloWorldService helloWorldService;

    @Test
    public void getAllDevelopers(){
        assertEquals(2, helloWorldService.getAllDevelopers().size());
    }





}